Move recursive_dict_update to satpy.utils
