+++ satpy/writers/__init__.py
@@ -35,7 +35,8 @@ except ImportError:
     from yaml import Loader as UnsafeLoader
 
 from satpy.config import (config_search_paths, glob_config,
-                          get_environ_config_dir, recursive_dict_update)
+                          get_environ_config_dir)
+from satpy.utils import recursive_dict_update
 from satpy import CHUNK_SIZE
 from satpy.plugin_base import Plugin
 from satpy.resample import get_area_def
